+++ mayan/apps/document_signatures/tests/test_settings.py
@@ -2,9 +2,9 @@ import importlib
 import logging
 
 from mayan.apps.document_signatures import storages
+from mayan.apps.smart_settings.utils import get_environment_variable_full_name
 from mayan.apps.storage.classes import DefinedStorage
 from mayan.apps.testing.tests.base import BaseTestCase
-from mayan.settings.literals import ENVIRONMENT_VARIABLE_PREFIX
 
 from ..literals import STORAGE_NAME_DOCUMENT_SIGNATURES_DETACHED_SIGNATURE
 from ..settings import setting_storage_backend_arguments
@@ -16,11 +16,11 @@ class SignatureStorageSettingsTestCase(BaseTestCase):
         importlib.reload(storages)
 
     def test_setting_storage_backend_arguments_invalid_value(self):
+        environment_variable_name = get_environment_variable_full_name(
+            name=setting_storage_backend_arguments.global_name
+        )
         self._set_environment_variable(
-            name='{}{}'.format(
-                ENVIRONMENT_VARIABLE_PREFIX,
-                setting_storage_backend_arguments.global_name
-            ), value='invalid_value'
+            name=environment_variable_name, value='invalid_value'
         )
 
         self.test_case_silenced_logger_new_level = logging.FATAL + 10