+++ mayan/apps/sources/tests/test_settings.py
@@ -2,10 +2,10 @@ import importlib
 import logging
 
 from mayan.apps.smart_settings.tests.mixins import TestMixinSettingDataType
+from mayan.apps.smart_settings.utils import get_environment_variable_full_name
 from mayan.apps.sources import storages
 from mayan.apps.storage.classes import DefinedStorage
 from mayan.apps.testing.tests.base import BaseTestCase
-from mayan.settings.literals import ENVIRONMENT_VARIABLE_PREFIX
 
 from ..literals import STORAGE_NAME_SOURCE_CACHE_FOLDER
 from ..settings import (
@@ -27,11 +27,11 @@ class SourcesStorageSettingsTestCase(BaseTestCase):
         importlib.reload(storages)
 
     def test_setting_source_cache_storage_backend_arguments_invalid_value(self):
+        environment_variable_name = get_environment_variable_full_name(
+            name=setting_source_cache_storage_backend_arguments.global_name
+        )
         self._set_environment_variable(
-            name='{}{}'.format(
-                ENVIRONMENT_VARIABLE_PREFIX,
-                setting_source_cache_storage_backend_arguments.global_name
-            ), value='invalid_value'
+            name=environment_variable_name, value='invalid_value'
         )
         self.test_case_silenced_logger_new_level = logging.FATAL + 10
         self._silence_logger(name='mayan.apps.storage.classes')