+++ mayan/apps/storage/tests/test_settings.py
@@ -2,9 +2,9 @@ import importlib
 import logging
 
 from mayan.apps.smart_settings.tests.mixins import TestMixinSettingDataType
+from mayan.apps.smart_settings.utils import get_environment_variable_full_name
 from mayan.apps.storage import storages
 from mayan.apps.testing.tests.base import BaseTestCase
-from mayan.settings.literals import ENVIRONMENT_VARIABLE_PREFIX
 
 from ..classes import DefinedStorage
 from ..literals import STORAGE_NAME_SHARED_UPLOADED_FILE
@@ -31,11 +31,11 @@ class StorageSettingsTestCase(TestMixinSettingDataType, BaseTestCase):
         self._do_test_setting_value_set()
 
     def test_setting_shared_storage_arguments_invalid_value(self):
+        environment_variable_name = get_environment_variable_full_name(
+            name=setting_shared_storage_arguments.global_name
+        )
         self._set_environment_variable(
-            name='{}{}'.format(
-                ENVIRONMENT_VARIABLE_PREFIX,
-                setting_shared_storage_arguments.global_name
-            ), value='invalid_value'
+            name=environment_variable_name, value='invalid_value'
         )
         self.test_case_silenced_logger_new_level = logging.FATAL + 10
         self._silence_logger(name='mayan.apps.storage.classes')